+++ docs/Changelog.md
@@ -2,13 +2,11 @@
 title: Roadmap
 description: Changelog and feature roadmap for Quartz Syncer.
 created: 2025-05-16T12:59:31Z+0200
-modified: 2025-05-18T12:41:44Z+0200
+modified: 2025-05-20T20:21:59Z+0200
 publish: true
 ---
 
 ## Upcoming
-
-*Awaiting Obsidian release feedback before working on next update.*
 
 ## Planned
 
@@ -23,6 +21,12 @@ publish: true
 - Manage Quartz plugins.
 
 ## Released
+
+### Version 1.5.3
+
+- Added note counts to the publication center.
+- Added tabbed settings menu.
+- Updated documentation.
 
 ### Version 1.5.2